# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "API documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/devices/fcm-token": {
            "post": {
                "tags": [
                    "UserDevice"
                ],
                "summary": "Auto generated endpoint",
                "operationId": "0ece5cfd8cb65c9ba5cd907448908b54",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "title": "RequestPayload",
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true,
                                            "deprecated": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Request processed successfully",
                                            "nullable": true,
                                            "deprecated": true
                                        },
                                        "data": {
                                            "title": "ResponseData",
                                            "type": "object",
                                            "deprecated": true,
                                            "additionalProperties": true
                                        },
                                        "meta": {
                                            "type": "object",
                                            "nullable": true,
                                            "deprecated": true,
                                            "additionalProperties": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/devices/expo-push-token": {
            "post": {
                "tags": [
                    "UserDevice"
                ],
                "summary": "Auto generated endpoint",
                "operationId": "4dc0dce7ae27ac55a21c9d8d30f0e363",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "title": "RequestPayload",
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Request processed successfully",
                                            "nullable": true
                                        },
                                        "data": {
                                            "title": "ResponseData",
                                            "type": "object",
                                            "additionalProperties": true
                                        },
                                        "meta": {
                                            "type": "object",
                                            "nullable": true,
                                            "additionalProperties": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/users/{userId}/public-profile": {
            "get": {
                "tags": [
                    "UserDiscovery"
                ],
                "summary": "Auto generated endpoint",
                "operationId": "e1392264d3b6f7b40a9d0a19a58183b2",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}